+++ manul/src/session/session.rs
@@ -130,7 +130,7 @@ pub enum RoundOutcome<P: Protocol, SP: SessionParameters> {
         /// The session object for the new round.
         session: Session<P, SP>,
         /// The messages intended for the new round cached during the previous round.
-        cached_messages: Vec<VerifiedMessage<SP>>,
+        cached_messages: Vec<VerifiedMessage<SP::Verifier>>,
     },
 }
 
@@ -231,7 +231,7 @@ where
         &self,
         rng: &mut impl CryptoRngCore,
         destination: &SP::Verifier,
-    ) -> Result<(Message, ProcessedArtifact<SP>), LocalError> {
+    ) -> Result<(Message<SP::Verifier>, ProcessedArtifact<SP>), LocalError> {
         let (direct_message, artifact) =
             self.round
                 .make_direct_message_with_artifact(rng, &self.serializer, destination)?;
@@ -241,6 +241,7 @@ where
             &self.signer,
             &self.session_id,
             self.round.id(),
+            destination,
             direct_message,
             self.echo_broadcast.clone(),
             self.normal_broadcast.clone(),
@@ -285,8 +286,8 @@ where
         &self,
         accum: &mut RoundAccumulator<P, SP>,
         from: &SP::Verifier,
-        message: Message,
-    ) -> Result<Option<VerifiedMessage<SP>>, LocalError> {
+        message: Message<SP::Verifier>,
+    ) -> Result<Option<VerifiedMessage<SP::Verifier>>, LocalError> {
         // Quick preliminary checks, before we proceed with more expensive verification
         let key = self.verifier();
         if self.transcript.is_banned(from) || accum.is_banned(from) {
@@ -381,7 +382,7 @@ where
     pub fn process_message(
         &self,
         rng: &mut impl CryptoRngCore,
-        message: VerifiedMessage<SP>,
+        message: VerifiedMessage<SP::Verifier>,
     ) -> ProcessedMessage<P, SP> {
         let processed = self.round.receive_message(
             rng,
@@ -544,7 +545,7 @@ pub struct RoundAccumulator<P: Protocol, SP: SessionParameters> {
     processing: BTreeSet<SP::Verifier>,
     payloads: BTreeMap<SP::Verifier, Payload>,
     artifacts: BTreeMap<SP::Verifier, Artifact>,
-    cached: BTreeMap<SP::Verifier, BTreeMap<RoundId, VerifiedMessage<SP>>>,
+    cached: BTreeMap<SP::Verifier, BTreeMap<RoundId, VerifiedMessage<SP::Verifier>>>,
     echo_broadcasts: BTreeMap<SP::Verifier, SignedMessagePart<EchoBroadcast>>,
     normal_broadcasts: BTreeMap<SP::Verifier, SignedMessagePart<NormalBroadcast>>,
     direct_messages: BTreeMap<SP::Verifier, SignedMessagePart<DirectMessage>>,
@@ -625,7 +626,7 @@ where
         }
     }
 
-    fn mark_processing(&mut self, message: &VerifiedMessage<SP>) -> Result<(), LocalError> {
+    fn mark_processing(&mut self, message: &VerifiedMessage<SP::Verifier>) -> Result<(), LocalError> {
         if !self.processing.insert(message.from().clone()) {
             Err(LocalError::new(format!(
                 "A message from {:?} is already marked as being processed",
@@ -732,7 +733,7 @@ where
         }
     }
 
-    fn cache_message(&mut self, message: VerifiedMessage<SP>) -> Result<(), LocalError> {
+    fn cache_message(&mut self, message: VerifiedMessage<SP::Verifier>) -> Result<(), LocalError> {
         let from = message.from().clone();
         let round_id = message.metadata().round_id();
         let cached = self.cached.entry(from.clone()).or_default();
@@ -754,14 +755,14 @@ pub struct ProcessedArtifact<SP: SessionParameters> {
 
 #[derive(Debug)]
 pub struct ProcessedMessage<P: Protocol, SP: SessionParameters> {
-    message: VerifiedMessage<SP>,
+    message: VerifiedMessage<SP::Verifier>,
     processed: Result<Payload, ReceiveError<SP::Verifier, P>>,
 }
 
-fn filter_messages<SP: SessionParameters>(
-    messages: BTreeMap<SP::Verifier, BTreeMap<RoundId, VerifiedMessage<SP>>>,
+fn filter_messages<Verifier>(
+    messages: BTreeMap<Verifier, BTreeMap<RoundId, VerifiedMessage<Verifier>>>,
     round_id: RoundId,
-) -> Vec<VerifiedMessage<SP>> {
+) -> Vec<VerifiedMessage<Verifier>> {
     messages
         .into_values()
         .filter_map(|mut messages| messages.remove(&round_id))
@@ -781,7 +782,7 @@ mod tests {
             Deserializer, DirectMessage, EchoBroadcast, NormalBroadcast, Protocol, ProtocolError,
             ProtocolValidationError, RoundId,
         },
-        testing::{BinaryFormat, TestSessionParams},
+        testing::{BinaryFormat, TestSessionParams, TestVerifier},
     };
 
     #[test]
@@ -835,9 +836,9 @@ mod tests {
         assert!(impls!(Session<DummyProtocol, SP>: Sync));
 
         // These objects are sent to/from message processing tasks
-        assert!(impls!(Message: Send));
+        assert!(impls!(Message<TestVerifier>: Send));
         assert!(impls!(ProcessedArtifact<SP>: Send));
-        assert!(impls!(VerifiedMessage<SP>: Send));
+        assert!(impls!(VerifiedMessage<TestVerifier>: Send));
         assert!(impls!(ProcessedMessage<DummyProtocol, SP>: Send));
     }
 }